requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-core
langchain-text-splitters
langchain-openai
langchain-chroma
langserve
sse_starlette